requirements.txt
Django==4.1.2
djangorestframework==3.14.0
pygments==2.13.0
static3==0.7.0
pytest-django==4.5.2
pytest==7.2.0
django-rest-knox==4.2.0
django-phonenumber-field[phonenumberslite]
Pillow==10.0.1
django-filter==23.1
psycopg2-binary==2.9.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
